Good Bye 2025
11:35 pm 27/12/2025 UTC +08
https://codeforces.com/contest/2178
I did the first question quite quickly, the trick was that each iteration the number of N's decreases by 1, so if there is <= 1 'Y', it can be done.
The second question took longer, but I realized that the first and last characters must be 's'. Then, I filled each 'uuu..uu' with 'usus...us'.